+++ exercises/patch-versioning/practice/src/main/java/versionworkflow/model/CustomerInfo.java
