+++ exercises/patch-versioning/practice/src/main/java/versionworkflow/model/SimpleCustomerMap.java
