+++ exercises/patch-versioning/solution/pom.xml
