+++ exercises/patch-versioning/solution/src/main/java/patchversion/Constants.java
@@ -1,4 +1,4 @@
-package versionworkflow;
+package patchversion;
 
 public class Constants {
