+++ exercises/patch-versioning/solution/src/main/java/patchversion/LoanProcessingWorker.java
@@ -1,4 +1,4 @@
-package versionworkflow;
+package patchversion;
 
 import io.temporal.client.WorkflowClient;
 import io.temporal.serviceclient.WorkflowServiceStubs;
