+++ exercises/patch-versioning/solution/src/main/java/patchversion/LoanProcessingWorkflow.java
@@ -1,9 +1,9 @@
-package versionworkflow;
+package patchversion;
 
 import io.temporal.workflow.WorkflowInterface;
 import io.temporal.workflow.WorkflowMethod;
 
-import versionworkflow.model.CustomerInfo;
+import patchversion.model.CustomerInfo;
 
 @WorkflowInterface
 public interface LoanProcessingWorkflow {
