+++ exercises/patch-versioning/solution/src/main/java/patchversion/Starter.java
@@ -1,7 +1,7 @@
-package versionworkflow;
+package patchversion;
 
-import versionworkflow.model.CustomerInfo;
-import versionworkflow.model.SimpleCustomerMap;
+import patchversion.model.CustomerInfo;
+import patchversion.model.SimpleCustomerMap;
 
 import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowOptions;
